requirements.txt
avro-python3==1.8.2
certifi==2020.6.20
chardet==3.0.4
confluent-kafka==1.5.0
fastavro==1.2.3
idna==2.10
redis==3.5.3
requests==2.24.0
urllib3==1.25.10
